requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113
torch==1.11.0+cu113
torchvision==0.12.0+cu113
numpy==1.22.1
opencv-python-headless>=4.2.0
scikit-image>=0.19.3
einops
tqdm
pillow
